requirements.txt
langchain==0.3.27
langchain_community==0.3.27
langchain_google_genai==2.1.8
langchain_pinecone==0.2.11
pinecone==7.3.0
python-dotenv==1.1.1
